requirements.txt
langchain_anthropic==0.2.1
langchain_core==0.3.6
langchain_openai==0.2.1
langgraph==0.2.28
openai==1.51.2
pydantic==2.9.2
python-dotenv==1.0.1
python_fasthtml==0.6.10
uvicorn==0.31.1
varname==0.13.5
langgraph.checkpoint.sqlite==1.0.4
